:recycle: move to home directory
